The suggested filename in the file save dialog is now checked for invalid filename characters (like ? or *).  Any found are replaced with a #.
Suggested filenames are produced from the item's label in the data manager.  Since the label may contain characters which are not valid in filenames (like a quotation mark), previously the suggested filename could have been invalid and so the file save dialog would open with no suggestion.  Now the user should always get a suggested filename.

Reviewed by Micah.