Fix hashCode/equals, refactor some
Got rid of NullConverter, replaced with a static factory on
ConverterImpl.  Refactored some methods for generics and good
method names.  Added a few utility classes for working with
Dictionaries, which OSGi uses a lot.

Still need to do some more small refactoring, and move the
ImmutableDictionary and friends to a utility bundle.